Refactored the current code models.
