+++ src/backend/KnowledgeNetwork.Domains.Code/Models/Enums/BasicBlockKind.cs
@@ -1,0 +1,27 @@
+namespace KnowledgeNetwork.Domains.Code.Models.Enums;
+
+/// <summary>
+/// Types of basic blocks in control flow
+/// </summary>
+public enum BasicBlockKind
+{
+    /// <summary>
+    /// Entry point of the method
+    /// </summary>
+    Entry,
+
+    /// <summary>
+    /// Exit point of the method
+    /// </summary>
+    Exit,
+
+    /// <summary>
+    /// Regular block with operations
+    /// </summary>
+    Block,
+
+    /// <summary>
+    /// Block that handles exceptions
+    /// </summary>
+    ExceptionHandler
+}
